use same function for combo and waterfall chart
