+++ frontend/src/metabase/visualizations/echarts/cartesian/waterfall/model/dataset.ts
@@ -8,7 +8,10 @@ import type {
   WaterfallXAxisModel,
   NumericAxisScaleTransforms,
 } from "metabase/visualizations/echarts/cartesian/model/types";
-import { replaceValues } from "metabase/visualizations/echarts/cartesian/model/dataset";
+import {
+  replaceValues,
+  replaceZeroesForLogScale,
+} from "metabase/visualizations/echarts/cartesian/model/dataset";
 import {
   WATERFALL_DATA_KEYS,
   WATERFALL_END_KEY,
@@ -16,44 +19,8 @@ import {
   WATERFALL_TOTAL_KEY,
   WATERFALL_VALUE_KEY,
 } from "metabase/visualizations/echarts/cartesian/waterfall/constants";
-import { isNotNull } from "metabase/lib/types";
 import { X_AXIS_DATA_KEY } from "metabase/visualizations/echarts/cartesian/constants/dataset";
 import { getNumberOr } from "metabase/visualizations/lib/settings/row-values";
-
-const replaceZerosForLogScale = (dataset: ChartDataset): ChartDataset => {
-  let hasZeros = false;
-  let minNonZeroValue = Infinity;
-
-  dataset.forEach(datum => {
-    const datumNumericValues = [
-      getNumberOr(datum[WATERFALL_START_KEY], null),
-      getNumberOr(datum[WATERFALL_END_KEY], null),
-    ].filter(isNotNull);
-
-    if (!hasZeros) {
-      hasZeros = datumNumericValues.includes(0);
-    }
-
-    minNonZeroValue = Math.min(
-      minNonZeroValue,
-      ...datumNumericValues.filter(number => number !== 0),
-    );
-  });
-
-  if (!hasZeros && minNonZeroValue > 0) {
-    return dataset;
-  }
-
-  if (minNonZeroValue < 0) {
-    throw Error(t`Y-axis must not cross 0 when using log scale.`);
-  }
-
-  const zeroReplacementValue = Math.min(minNonZeroValue, 1);
-
-  return replaceValues(dataset, (dataKey: DataKey, value: RowValue) =>
-    dataKey !== X_AXIS_DATA_KEY && value === 0 ? zeroReplacementValue : value,
-  );
-};
 
 export const getWaterfallDataset = (
   dataset: ChartDataset,
@@ -103,7 +70,10 @@ export const getWaterfallDataset = (
   }
 
   if (settings["graph.y_axis.scale"] === "log") {
-    transformedDataset = replaceZerosForLogScale(transformedDataset);
+    transformedDataset = replaceZeroesForLogScale(
+      transformedDataset,
+      WATERFALL_DATA_KEYS,
+    );
   }
 
   return replaceValues(
